backend/requirements.txt
# Web Server & API
fastapi
uvicorn
python-multipart

# AI Orchestration (The Brain)
langgraph
langchain
langchain-core
langchain-community
langchain-openai
langchain-text-splitters

# Local Embeddings (Privacy-First)
sentence-transformers
langchain-huggingface

# Database & Data Handling
qdrant-client
pydantic
python-dotenv
pypdf